changed temp/humidity rounding to 1 d.p and moved gc.collect()
dht22 seems to record values to 1 d.p. so I changed the code to reflect this. Also changed the gc.collect() method yet again, removed the garbageCore() and moved the code to the temperatureCore(). There was an error because due to a core not closing down so when I tried opening a new core it failed and caused a crash. So instead I moved the garbage collection to the temperatureCore() and am controlling it with a global boolean variable which toggles every log interval period and resets once garbage has been collected. This prevents attempting to use more than 2 cores at the same time. I haven't tested it extensively so I'm not sure if the delay caused by gc.collect() and dht_sensor.measure() combined is greater than the time it takes for the main loop to cycle but it seems fine so far in early stages of running