added an indicator for when the system is saving
when saving, the system now blinks the onboard led to indicate that the system is saving to the mircoSD card. I think the operation will be too fast for the led to even turn on but in the case that the system hangs this should warn the user not to power down the system. Again from prior testing I severely doubt the led will even turn on but its a feature now I guess...!